+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Current (in progress)
 
 - Fix community resource creation and display [#1733](https://github.com/opendatateam/udata/pull/1733)
+- Failsafe JS cache storage: use a custom in-memory storage as fallback when access to `sessionStorage` is not allowed [#1742](https://github.com/opendatateam/udata/pull/1742)
 
 ## 1.4.0 (2018-06-06)